--- a/flair/trainers/language_model_trainer.py
+++ b/flair/trainers/language_model_trainer.py
@@ -1,5 +1,6 @@
 import time, datetime
 import random
+import sys
 import logging
 from pathlib import Path
 from typing import Union
@@ -7,6 +8,11 @@ from typing import Union
 from torch import cuda
 from torch.utils.data import Dataset, DataLoader
 from torch.optim.sgd import SGD
+
+try:
+    from apex import amp
+except ImportError:
+    amp = None
 
 import flair
 from flair.data import Dictionary
@@ -248,9 +254,18 @@ class LanguageModelTrainer:
         max_epochs: int = 1000,
         checkpoint: bool = False,
         grow_to_sequence_length: int = 0,
+        apex: bool = False,
+        apex_opt_level: str = 'O1',
         **kwargs,
     ):
 
+        if apex:
+            if sys.version_info < (3, 0):
+                raise RuntimeError("Apex currently only supports Python 3. Aborting.")
+            if amp is None:
+                raise RuntimeError("Failed to import apex. Please install apex from https://www.github.com/nvidia/apex "
+                                   "to enable mixed-precision training.")
+                
         # cast string to Path
         if type(base_path) is str:
             base_path = Path(base_path)
@@ -283,6 +298,11 @@ class LanguageModelTrainer:
                     optimizer, verbose=True, factor=anneal_factor, patience=patience
                 )
 
+            if apex:
+                self.model, optimizer = amp.initialize(self.model, optimizer,
+                                                       opt_level=apex_opt_level
+                                                       )
+            
             training_generator = DataLoader(
                 self.corpus.train, shuffle=False, num_workers=self.num_workers
             )
@@ -356,7 +376,12 @@ class LanguageModelTrainer:
 
                         # try to predict the targets
                         loss = self.loss_function(output.view(-1, ntokens), targets)
-                        loss.backward()
+                                            # Backward
+                        if apex:
+                            with amp.scale_loss(loss, optimizer) as scaled_loss:
+                                scaled_loss.backward()
+                        else:
+                            loss.backward()
 
                         # `clip_grad_norm` helps prevent the exploding gradient problem in RNNs / LSTMs.
                         torch.nn.utils.clip_grad_norm_(self.model.parameters(), clip)