@alanakbik 2019-07-29
@vinhngx This is great, thanks for adding this! Could you also include APEX support for the `LanguageModelTrainer` class in this PR? Training LMs is currently the most resource-intensive task so it would be interesting to see what kind of speed gains we get here.