@vinhngx 2019-07-30
Thanks @alanakbik. I've added APEX to the language model trainer.
Mixed precision training can be enabled with:

```
# train your language model
trainer = LanguageModelTrainer(language_model, corpus)

trainer.train('resources/taggers/language_model',
              sequence_length=256,
              mini_batch_size=256,
              max_epochs=10,
              apex=True)
```