@stefan-it 2019-07-30
I tried it in combination with `BertEmbeddings` on a V-100 GPU:

Without `apex`:

```bash
2019-07-29 20:42:27,531 epoch 1 - iter 0/795 - loss 55.33607864 throughput (samples/sec): 2900.75
2019-07-29 20:43:08,202 epoch 1 - iter 79/795 - loss 9.06459931 throughput (samples/sec): 31.19
2019-07-29 20:43:50,289 epoch 1 - iter 158/795 - loss 7.42229333 throughput (samples/sec): 30.12
2019-07-29 20:44:31,381 epoch 1 - iter 237/795 - loss 6.46287990 throughput (samples/sec): 30.85
2019-07-29 20:45:12,541 epoch 1 - iter 316/795 - loss 5.83835078 throughput (samples/sec): 30.80
2019-07-29 20:45:53,771 epoch 1 - iter 395/795 - loss 5.42233314 throughput (samples/sec): 30.75
2019-07-29 20:46:34,585 epoch 1 - iter 474/795 - loss 5.05916398 throughput (samples/sec): 31.06
2019-07-29 20:47:16,937 epoch 1 - iter 553/795 - loss 4.78533588 throughput (samples/sec): 29.93
2019-07-29 20:47:59,127 epoch 1 - iter 632/795 - loss 4.53341620 throughput (samples/sec): 30.05
2019-07-29 20:48:40,761 epoch 1 - iter 711/795 - loss 4.35730412 throughput (samples/sec): 30.45
2019-07-29 20:49:22,217 epoch 1 - iter 790/795 - loss 4.17682579 throughput (samples/sec): 30.58
2019-07-29 20:49:23,789 ----------------------------------------------------------------------------------------------------
2019-07-29 20:49:23,789 EPOCH 1 done: loss 4.1919 - lr 0.1000
2019-07-29 20:51:24,836 DEV : loss 2.9106078147888184 - score 0.6502
2019-07-29 20:51:24,872 BAD EPOCHS (no improvement): 0
2019-07-29 20:51:25,456 ----------------------------------------------------------------------------------------------------
2019-07-29 20:51:25,563 epoch 2 - iter 0/795 - loss 0.98591888 throughput (samples/sec): 12115.87
2019-07-29 20:51:35,162 epoch 2 - iter 79/795 - loss 2.51360194 throughput (samples/sec): 133.39
2019-07-29 20:51:44,908 epoch 2 - iter 158/795 - loss 2.54102277 throughput (samples/sec): 131.35
2019-07-29 20:51:54,648 epoch 2 - iter 237/795 - loss 2.55147406 throughput (samples/sec): 131.44
2019-07-29 20:52:04,199 epoch 2 - iter 316/795 - loss 2.45608709 throughput (samples/sec): 134.08
2019-07-29 20:52:13,655 epoch 2 - iter 395/795 - loss 2.43405817 throughput (samples/sec): 135.43
2019-07-29 20:52:23,213 epoch 2 - iter 474/795 - loss 2.42034678 throughput (samples/sec): 133.95
2019-07-29 20:52:33,684 epoch 2 - iter 553/795 - loss 2.36586307 throughput (samples/sec): 122.13
2019-07-29 20:52:44,983 epoch 2 - iter 632/795 - loss 2.32865828 throughput (samples/sec): 113.07
2019-07-29 20:52:55,985 epoch 2 - iter 711/795 - loss 2.28737589 throughput (samples/sec): 116.18
2019-07-29 20:53:07,091 epoch 2 - iter 790/795 - loss 2.28597844 throughput (samples/sec): 115.07
2019-07-29 20:53:07,747 ----------------------------------------------------------------------------------------------------
```

With `apex` enabled:

```bash
2019-07-29 23:40:03,869 epoch 1 - iter 0/795 - loss 40.28230286 throughput (samples/sec): 2979.66
2019-07-29 23:40:49,085 epoch 1 - iter 79/795 - loss 10.62385516 throughput (samples/sec): 28.04
2019-07-29 23:41:32,755 epoch 1 - iter 158/795 - loss 7.94562467 throughput (samples/sec): 29.03
2019-07-29 23:42:15,033 epoch 1 - iter 237/795 - loss 6.78094378 throughput (samples/sec): 29.98
2019-07-29 23:42:59,963 epoch 1 - iter 316/795 - loss 6.05153041 throughput (samples/sec): 28.21
2019-07-29 23:43:42,590 epoch 1 - iter 395/795 - loss 5.49845262 throughput (samples/sec): 29.74
2019-07-29 23:44:25,641 epoch 1 - iter 474/795 - loss 5.16456086 throughput (samples/sec): 29.48
2019-07-29 23:45:09,427 epoch 1 - iter 553/795 - loss 4.82567009 throughput (samples/sec): 28.95
2019-07-29 23:45:52,538 epoch 1 - iter 632/795 - loss 4.57211298 throughput (samples/sec): 29.40
2019-07-29 23:46:33,286 epoch 1 - iter 711/795 - loss 4.38901180 throughput (samples/sec): 31.11
2019-07-29 23:47:17,371 epoch 1 - iter 790/795 - loss 4.20550554 throughput (samples/sec): 28.75
2019-07-29 23:47:19,562 ----------------------------------------------------------------------------------------------------
2019-07-29 23:47:19,562 EPOCH 1 done: loss 4.2002 - lr 0.1000
2019-07-29 23:49:26,519 DEV : loss 3.2141149044036865 - score 0.5982
2019-07-29 23:49:26,558 BAD EPOCHS (no improvement): 0
2019-07-29 23:49:27,139 ----------------------------------------------------------------------------------------------------
2019-07-29 23:49:27,282 epoch 2 - iter 0/795 - loss 1.12042511 throughput (samples/sec): 8974.31
2019-07-29 23:49:37,825 epoch 2 - iter 79/795 - loss 2.54959387 throughput (samples/sec): 121.26
2019-07-29 23:49:48,250 epoch 2 - iter 158/795 - loss 2.49878981 throughput (samples/sec): 122.71
2019-07-29 23:49:58,878 epoch 2 - iter 237/795 - loss 2.43002224 throughput (samples/sec): 120.30
2019-07-29 23:50:09,504 epoch 2 - iter 316/795 - loss 2.40728975 throughput (samples/sec): 120.34
2019-07-29 23:50:20,038 epoch 2 - iter 395/795 - loss 2.34507697 throughput (samples/sec): 121.37
2019-07-29 23:50:30,534 epoch 2 - iter 474/795 - loss 2.32682362 throughput (samples/sec): 121.85
2019-07-29 23:50:41,203 epoch 2 - iter 553/795 - loss 2.30398093 throughput (samples/sec): 119.88
2019-07-29 23:50:51,679 epoch 2 - iter 632/795 - loss 2.27872584 throughput (samples/sec): 122.10
2019-07-29 23:51:02,250 epoch 2 - iter 711/795 - loss 2.25251733 throughput (samples/sec): 120.96
2019-07-29 23:51:13,189 epoch 2 - iter 790/795 - loss 2.22839135 throughput (samples/sec): 116.84
2019-07-29 23:51:13,751 ----------------------------------------------------------------------------------------------------
```

`samples/sec` are always lower with `apex` enabled.