@alanakbik 2019-07-30
@stefan-it that is strange. Do you have a setup to test a language model training task? LM training is the only task for which we currently have 100% GPU usage.