@vinhngx 2019-07-30
@stefan-it There is some overhead incurred with mixed precision training, so models with low arithmetic intensity, i.e. the ratio of algorithm implementation operations and the number of bytes accessed, might not necessarily benefit. 

For NVIDIA GPUs with Tensor cores, as a general performance guide, dimensions (such as batch size, input size, output size, and channel counts) should be powers of two if under 256, or 
otherwise divisible by 8 if above 256.  Language models with odd vocabulary sizes should be therefore padded to a multiple of 8. For more information, check out the
[NVIDIA Deep Learning Performance Guide](
  https://docs.nvidia.com/deeplearning/sdk/dl-performance-guide/index.html).

Large models such as BERT pre-training have seen up to 4x speed up. 
https://github.com/NVIDIA/DeepLearningExamples/tree/master/PyTorch/LanguageModeling/BERT

I'm curious to see which Flair models will see most benefits :)