@alanakbik 2019-07-31
I've tested APEX for training a character-LM over the 1-B word corpus. It's a 512 state, 1-layer RNN with a dictionary size of 280, sequence length of 256 and a mini-batch size of 256. 

Without APEX: 

```console
| end of split   1 / 99 | epoch   1 | time: 54.22s | valid loss  1.61 | valid ppl     5.01 | learning rate 20.0000
| end of split   2 / 99 | epoch   1 | time: 53.45s | valid loss  1.40 | valid ppl     4.06 | learning rate 20.0000
| end of split   3 / 99 | epoch   1 | time: 53.72s | valid loss  1.54 | valid ppl     4.68 | learning rate 20.0000
| end of split   4 / 99 | epoch   1 | time: 53.24s | valid loss  1.29 | valid ppl     3.63 | learning rate 20.0000
| end of split   5 / 99 | epoch   1 | time: 53.53s | valid loss  1.26 | valid ppl     3.52 | learning rate 20.0000
| end of split   6 / 99 | epoch   1 | time: 53.71s | valid loss  1.24 | valid ppl     3.46 | learning rate 20.0000
| end of split   7 / 99 | epoch   1 | time: 53.62s | valid loss  1.23 | valid ppl     3.41 | learning rate 20.0000
| end of split   8 / 99 | epoch   1 | time: 53.76s | valid loss  1.21 | valid ppl     3.36 | learning rate 20.0000
| end of split   9 / 99 | epoch   1 | time: 52.98s | valid loss  1.21 | valid ppl     3.37 | learning rate 20.0000
| end of split  10 / 99 | epoch   1 | time: 53.71s | valid loss  1.19 | valid ppl     3.30 | learning rate 20.0000
| end of split  11 / 99 | epoch   1 | time: 53.77s | valid loss  1.21 | valid ppl     3.36 | learning rate 20.0000
| end of split  12 / 99 | epoch   1 | time: 53.75s | valid loss  1.19 | valid ppl     3.28 | learning rate 20.0000
| end of split  13 / 99 | epoch   1 | time: 53.85s | valid loss  1.18 | valid ppl     3.24 | learning rate 20.0000
| end of split  14 / 99 | epoch   1 | time: 53.36s | valid loss  1.17 | valid ppl     3.22 | learning rate 20.0000
| end of split  15 / 99 | epoch   1 | time: 53.76s | valid loss  1.17 | valid ppl     3.21 | learning rate 20.0000
| end of split  16 / 99 | epoch   1 | time: 53.79s | valid loss  1.17 | valid ppl     3.21 | learning rate 20.0000
| end of split  17 / 99 | epoch   1 | time: 53.82s | valid loss  1.18 | valid ppl     3.27 | learning rate 20.0000
| end of split  18 / 99 | epoch   1 | time: 53.87s | valid loss  1.16 | valid ppl     3.18 | learning rate 20.0000
| end of split  19 / 99 | epoch   1 | time: 53.76s | valid loss  1.16 | valid ppl     3.20 | learning rate 20.0000
| end of split  20 / 99 | epoch   1 | time: 54.16s | valid loss  1.16 | valid ppl     3.18 | learning rate 20.0000
| end of split  21 / 99 | epoch   1 | time: 53.21s | valid loss  1.15 | valid ppl     3.17 | learning rate 20.0000
| end of split  22 / 99 | epoch   1 | time: 53.66s | valid loss  1.15 | valid ppl     3.14 | learning rate 20.0000
| end of split  23 / 99 | epoch   1 | time: 53.96s | valid loss  1.15 | valid ppl     3.15 | learning rate 20.0000
| end of split  24 / 99 | epoch   1 | time: 53.62s | valid loss  1.17 | valid ppl     3.23 | learning rate 20.0000
| end of split  25 / 99 | epoch   1 | time: 53.81s | valid loss  1.14 | valid ppl     3.12 | learning rate 20.0000
| end of split  26 / 99 | epoch   1 | time: 53.67s | valid loss  1.14 | valid ppl     3.12 | learning rate 20.0000
| end of split  27 / 99 | epoch   1 | time: 53.88s | valid loss  1.14 | valid ppl     3.12 | learning rate 20.0000
| end of split  28 / 99 | epoch   1 | time: 53.54s | valid loss  1.15 | valid ppl     3.17 | learning rate 20.0000
| end of split  29 / 99 | epoch   1 | time: 53.41s | valid loss  1.13 | valid ppl     3.10 | learning rate 20.0000
| end of split  30 / 99 | epoch   1 | time: 53.80s | valid loss  1.13 | valid ppl     3.10 | learning rate 20.0000
```

With APEX: 

```console
| end of split   1 / 99 | epoch   1 | time: 35.92s | valid loss  1.63 | valid ppl     5.12 | learning rate 20.0000
| end of split   2 / 99 | epoch   1 | time: 36.27s | valid loss  1.41 | valid ppl     4.09 | learning rate 20.0000
| end of split   3 / 99 | epoch   1 | time: 34.98s | valid loss  1.56 | valid ppl     4.75 | learning rate 20.0000
| end of split   4 / 99 | epoch   1 | time: 35.17s | valid loss  1.28 | valid ppl     3.61 | learning rate 20.0000
| end of split   5 / 99 | epoch   1 | time: 35.14s | valid loss  1.26 | valid ppl     3.53 | learning rate 20.0000
| end of split   6 / 99 | epoch   1 | time: 35.93s | valid loss  1.24 | valid ppl     3.45 | learning rate 20.0000
| end of split   7 / 99 | epoch   1 | time: 35.41s | valid loss  1.22 | valid ppl     3.40 | learning rate 20.0000
| end of split   8 / 99 | epoch   1 | time: 34.83s | valid loss  1.22 | valid ppl     3.38 | learning rate 20.0000
| end of split   9 / 99 | epoch   1 | time: 35.57s | valid loss  1.21 | valid ppl     3.36 | learning rate 20.0000
| end of split  10 / 99 | epoch   1 | time: 36.25s | valid loss  1.20 | valid ppl     3.31 | learning rate 20.0000
| end of split  11 / 99 | epoch   1 | time: 36.07s | valid loss  1.22 | valid ppl     3.38 | learning rate 20.0000
| end of split  12 / 99 | epoch   1 | time: 35.28s | valid loss  1.19 | valid ppl     3.30 | learning rate 20.0000
| end of split  13 / 99 | epoch   1 | time: 35.93s | valid loss  1.18 | valid ppl     3.26 | learning rate 20.0000
| end of split  14 / 99 | epoch   1 | time: 35.78s | valid loss  1.17 | valid ppl     3.23 | learning rate 20.0000
| end of split  15 / 99 | epoch   1 | time: 34.98s | valid loss  1.17 | valid ppl     3.22 | learning rate 20.0000
| end of split  16 / 99 | epoch   1 | time: 36.06s | valid loss  1.17 | valid ppl     3.22 | learning rate 20.0000
| end of split  17 / 99 | epoch   1 | time: 35.85s | valid loss  1.19 | valid ppl     3.28 | learning rate 20.0000
| end of split  18 / 99 | epoch   1 | time: 35.33s | valid loss  1.16 | valid ppl     3.19 | learning rate 20.0000
| end of split  19 / 99 | epoch   1 | time: 34.77s | valid loss  1.16 | valid ppl     3.20 | learning rate 20.0000
| end of split  20 / 99 | epoch   1 | time: 36.08s | valid loss  1.16 | valid ppl     3.19 | learning rate 20.0000
| end of split  21 / 99 | epoch   1 | time: 35.72s | valid loss  1.15 | valid ppl     3.17 | learning rate 20.0000
| end of split  22 / 99 | epoch   1 | time: 35.86s | valid loss  1.15 | valid ppl     3.16 | learning rate 20.0000
| end of split  23 / 99 | epoch   1 | time: 36.13s | valid loss  1.15 | valid ppl     3.16 | learning rate 20.0000
| end of split  24 / 99 | epoch   1 | time: 35.09s | valid loss  1.18 | valid ppl     3.27 | learning rate 20.0000
| end of split  25 / 99 | epoch   1 | time: 36.16s | valid loss  1.15 | valid ppl     3.15 | learning rate 20.0000
| end of split  26 / 99 | epoch   1 | time: 34.85s | valid loss  1.14 | valid ppl     3.13 | learning rate 20.0000
| end of split  27 / 99 | epoch   1 | time: 35.22s | valid loss  1.14 | valid ppl     3.14 | learning rate 20.0000
| end of split  28 / 99 | epoch   1 | time: 35.74s | valid loss  1.15 | valid ppl     3.17 | learning rate 20.0000
| end of split  29 / 99 | epoch   1 | time: 35.24s | valid loss  1.14 | valid ppl     3.11 | learning rate 20.0000
| end of split  30 / 99 | epoch   1 | time: 36.01s | valid loss  1.14 | valid ppl     3.12 | learning rate 20.0000
```

So we are seeing average time per split reduced from ~ 54 seconds to ~ 35 seconds, which is a nice speed up!

I previously tested a smaller language model with only 128 states and there was no difference. I could image that the difference is more pronounced if we train larger LMs. Our typical size LM has 2048 states, so maybe there's even better increases here.